Force log-white listed logs
