+++ src/Azure.Functions.Cli/Common/Utilities.cs
@@ -2,6 +2,9 @@
 using Colors.Net;
 using Colors.Net.StringColorExtensions;
 using Microsoft.Azure.WebJobs.Script;
+using Microsoft.Extensions.Logging;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using System;
 using System.Diagnostics;
 using System.IO;
@@ -163,6 +166,20 @@ namespace Azure.Functions.Cli
             var localPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "azure-functions-core-tools");
             FileSystemHelpers.EnsureDirectory(localPath);
             return localPath;
+        }
+
+        internal static LogLevel GetHostJsonDefaultLogLevel()
+        {
+            try
+            {
+                var hostJson = JsonConvert.DeserializeObject<JObject>(FileSystemHelpers.ReadAllTextFromFile("host.json"));
+                return (LogLevel)Enum.Parse(typeof(LogLevel), hostJson["Logging"]["LogLevel"]["Default"].ToString());
+            }
+            catch (Exception)
+            {
+                // Default log level
+                return LogLevel.Information;
+            }
         }
     }
 }
